[pallas] Add lowering errors for block shapes that are not supported.
Previously these errors came Mosaic with less useful stack traces, and in the case of GPU we get a crash instead of an exception.

In order to implement this I had to make sure that _convert_block_spec_to_block_mapping always returns a BlockMapping even if the block spec is no_block_spec.

PiperOrigin-RevId: 653886179